shaved-vs-unshaved-flask-application/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
applaunchservices==0.2.1
appnope==0.1.0
appscript==1.0.1
argh==0.26.2
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0.1.post1
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.4
bokeh==2.0.1
boto==2.49.0
Bottleneck==1.3.2
branca==0.3.1
cachetools==4.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.4.0
clyent==1.2.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.8
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
convertdate==2.2.0
cryptography==2.8
cvxpy==1.0.26
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dash==1.3.0
dash-core-components==1.2.0
dash-daq==0.2.1
dash-html-components==1.0.1
dash-renderer==1.1.0
dash-table==4.3.0
dask==2.15.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
distributed==2.15.0
Django==3.0.5
docutils==0.16
ecos==2.0.7.post1
entrypoints==0.3
ephem==3.7.7.0
et-xmlfile==1.0.1
fastcache==1.1.0
fbprophet==0.5
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-Compress==1.4.0
Flask-Uploads==0.2.1
folium==0.10.0
fsspec==0.7.1
future==0.18.2
gast==0.2.2
gevent==1.4.0
glob2==0.7
gmplot==1.2.0
gmpy2==2.0.8
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
gym==0.17.1
h5py==2.10.0
HeapDict==1.0.1
holidays==0.10.1
html5lib==1.0.1
hypothesis==5.8.3
idna==2.9
image==1.5.31
imageio==2.8.0
imagesize==1.2.0
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==1.5.0
inflection==0.3.1
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
lief==0.9.0
llvmlite==0.32.0
locket==0.2.0
lunardate==0.2.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
multiprocess==0.70.9
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.49.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.2.0
osqp==0.6.1
packaging==20.3
pandas==1.0.3
pandasql==0.7.3
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
plotly==4.1.1
pluggy==0.13.1
ply==3.11
pmdarima==1.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.1.1
pyglet==1.5.0
Pygments==2.6.1
pylint==2.4.4
PyMeeus==0.3.6
pymongo==3.9.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pyspark==2.4.5
pystan==2.19.1.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-kucoin==2.1.2
python-language-server==0.31.10
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.3
QtPy==1.9.0
Quandl==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.0
Rtree==0.9.3
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scikit-surprise==1.1.0
scipy==1.4.1
scs==2.1.1.post2
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
setuptools-git==1.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.2
spyder-kernels==1.9.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.18.0
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==2.0.0
zipp==2.2.0
